requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
gpsd-py3==0.3.0
idna==3.3
requests==2.27.1
urllib3==1.26.8
